Move best bet helper methods to helper module
So they can be used by other tests.
